+++ compiler/ml/builtin_attributes.mli
@@ -93,7 +93,5 @@ val explicit_arity : Parsetree.attributes -> bool
 
 val immediate : Parsetree.attributes -> bool
 
-val has_allow_mutation : Parsetree.attributes -> bool
-
 val has_unboxed : Parsetree.attributes -> bool
 val has_boxed : Parsetree.attributes -> bool